requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
pyjwt==2.8.0
slowapi==0.1.9
python-multipart==0.0.6